refactor: Use bit to import design system
