+++ package.json
@@ -11,6 +11,11 @@
     "npm": "6.9.0"
   },
   "dependencies": {
+    "@bit/opencollective.design-system.components.styled-container": "1.0.0",
+    "@bit/opencollective.design-system.components.styled-hr": "1.0.0",
+    "@bit/opencollective.design-system.components.styled-link": "1.1.0",
+    "@bit/opencollective.design-system.components.styled-text": "1.0.0",
+    "@bit/opencollective.design-system.theme": "1.0.0",
     "@styled-system/theme-get": "5.1.2",
     "clean-tag": "~2.0.3",
     "cors": "~2.8.5",
