+++ pages/_app.js
@@ -2,8 +2,7 @@ import React from 'react';
 import App from 'next/app';
 import { ThemeProvider, createGlobalStyle } from 'styled-components';
 import { IntlProvider } from 'react-intl';
-
-import theme from '../lib/constants/theme';
+import theme from '../lib/theme';
 
 const GlobalStyles = createGlobalStyle`
   @font-face {
